add numbers to text and escape numbers
