+++ ave/game_loader.py
@@ -1,52 +1,11 @@
 import re
-import os
 import json
 import urllib.request
 from .game import Game, Room, Item, Number
 from .game import (TextWithRequirements, OptionWithRequirements,
                    NameWithRequirements)
-from .ave_format import symbols, attributes
-
-
-def between(text, pre, post):
-    if pre not in text or post not in text:
-        return ""
-    else:
-        return text.split(pre, 1)[1].split(post, 1)[0]
-
-
-def _replacements(string):
-    with open(os.path.join(os.path.dirname(os.path.realpath(__file__)),
-                           "../VERSION")) as f:
-        v = f.read().strip()
-    string = v.join(string.split("%v%"))
-    return string
-
-
-def remove_links(txt):
-    return re.sub(r"\[(.*)\]\((.*)\)", r"\2", txt)
-
-
-def clean(string):
-    string = string.replace("<newline>", "\n")
-    return _replacements(string.strip())
-
-
-def _escape(matches):
-    text = matches[1]
-    for i, j in symbols.items():
-        text = text.replace(i, j)
-    return text
-
-
-def escape(line):
-    return re.sub(r"<\|(.*?)\|>", _escape, line)
-
-
-def unescape(text):
-    for i, j in symbols.items():
-        text = text.replace(j, i)
-    return text
+from .ave_format import attributes
+from .escaping import escape, unescape, clean, between
 
 
 def parse_requirements(req, id_of_text="text"):
